feat: add setup script for Decap CMS GitHub Backend with Firebase Functions
- Create setup.sh script to automate the setup process
- Install necessary dependencies for the project and functions
- Prompt user for Firebase initialization and GitHub OAuth credentials
- Set Firebase config values for GitHub integration
- Provide instructions for starting the local server and deploying to Firebase
