--- .editorconfig
@@ -1,20 +1,0 @@
-# EditorConfig is awesome: https://EditorConfig.org
-
-root = true
-
-[*]
-indent_style = space
-indent_size = 2
-end_of_line = lf
-charset = utf-8
-trim_trailing_whitespace = true
-insert_final_newline = true
-max_line_length = 140
-
-[*.{js,ts}]
-block_comment_start = /**
-block_comment = *
-block_comment_end = */
-
-[*.md]
-trim_trailing_whitespace = false
