--- .prettierignore
@@ -1,3 +1,0 @@
-node_modules
-CHANGELOG.md
-dist
