--- CHANGELOG.md
@@ -1,39 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
-
-### [1.2.4](https://github.com/njfamirm/decap-cms-github-backend/compare/v1.2.3...v1.2.4) (2025-05-02)
-
-### [1.2.3](https://github.com/njfamirm/decap-cms-github-backend/compare/v1.2.2...v1.2.3) (2025-03-02)
-
-### [1.2.2](https://github.com/njfamirm/decap-cms-github-backend/compare/v1.2.0...v1.2.2) (2025-03-02)
-
-### [1.2.1](https://github.com/njfamirm/decap-cms-github-backend/compare/v1.2.0...v1.2.1) (2025-03-02)
-
-## [1.2.0](https://github.com/njfamirm/decap-cms-github-backend/compare/v1.1.1...v1.2.0) (2025-01-10)
-
-
-### Features
-
-* add demo ([#29](https://github.com/njfamirm/decap-cms-github-backend/issues/29)) ([0664dec](https://github.com/njfamirm/decap-cms-github-backend/commit/0664deca159bdbddd955a1de5f0076c5477d559c))
-
-### [1.1.1](https://github.com/njfamirm/decap-cms-github-backend/compare/v1.1.0...v1.1.1) (2025-01-10)
-
-
-### Bug Fixes
-
-* install correct deps ([9d98ecc](https://github.com/njfamirm/decap-cms-github-backend/commit/9d98ecce5fdc2b49167c68e9e64e537e1ebe94df))
-
-## [1.1.0](https://github.com/njfamirm/decap-cms-github-backend/compare/v1.0.0...v1.1.0) (2025-01-10)
-
-
-### Features
-
-* fix scope ([#8](https://github.com/njfamirm/decap-cms-github-backend/issues/8)) ([a2fa589](https://github.com/njfamirm/decap-cms-github-backend/commit/a2fa589a5ce7a5ea5a52826eb85c0224aa0ecf0a))
-* set scope with env ([2ae80e3](https://github.com/njfamirm/decap-cms-github-backend/commit/2ae80e3ee4e56963f6a3af5871bc6b8c0840ffac))
-* some enhancement ([#26](https://github.com/njfamirm/decap-cms-github-backend/issues/26)) ([f14a6fb](https://github.com/njfamirm/decap-cms-github-backend/commit/f14a6fbf61df123b20cdf13654208d7f2c0f1c5c))
-
-
-### Bug Fixes
-
-* remove scope user ([0cdc3fa](https://github.com/njfamirm/decap-cms-github-backend/commit/0cdc3fa521a6c7183b8771440d3d9defd2ace101))
