--- demo/.env
@@ -1,8 +1,0 @@
-CMS_BACKEND_IMAGE='ghcr.io/njfamirm/decap-cms-github-backend:1.2.3'
-OAUTH_GITHUB_CLIENT_ID='maybe-secret'
-OAUTH_GITHUB_CLIENT_SECRET='secret'
-DOMAIN='decap-demo.njfamirm.ir'
-OAUTH_GITHUB_SCOPE='public_repo'
-
-# Enable debug logging
-CMS_BACKEND_DEBUG=1
