--- demo/admin/admin.js
@@ -1,3 +1,0 @@
-import CMS from 'netlify-cms-app';
-
-CMS.init();
