--- demo/admin/index.html
@@ -1,12 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta name="robots" content="noindex" />
-    <title>Website | Admin Panel</title>
-    <script type="module" src="/admin/admin.min.js"></script>
-    <link href="/config.yaml" type="text/yaml" rel="cms-config-url" />
-  </head>
-  <body></body>
-</html>
