--- demo/config.yaml
@@ -1,21 +1,0 @@
-backend:
-  name: github
-  branch: main
-  repo: njfamirm/decap-cms-github-backend
-  base_url: https://decap-demo.njfamirm.ir/
-
-publish_mode: editorial_workflow
-
-media_folder: 'static/uploads'
-public_folder: '/uploads'
-
-collections:
-  - name: blog
-    label: Blog
-    folder: 'content/blog'
-    create: true
-    slug: '{{year}}-{{month}}-{{day}}-{{slug}}'
-    fields:
-      - { label: 'Title', name: 'title', widget: 'string' }
-      - { label: 'Date', name: 'date', widget: 'datetime' }
-      - { label: 'Body', name: 'body', widget: 'markdown' }
