--- demo/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@njfamirm/decap-cms-github-backend-demo",
-  "private": true,
-  "description": "Demo of self-hosted decap cms github backend.",
-  "license": "MIT",
-  "author": "S. Amir Mohammad Najafi <njfamirm@gamil.com> (https://www.njfamirm.ir)",
-  "type": "module",
-  "scripts": {
-    "build": "esbuild demo/admin.js --bundle --minify --outfile=demo/admin.min.js",
-    "serve": "wds"
-  },
-  "dependencies": {
-    "netlify-cms-app": "^2.15.72"
-  },
-  "devDependencies": {
-    "@types/node": "^22.10.5",
-    "@web/dev-server": "^0.4.6",
-    "esbuild": "^0.25.0"
-  },
-  "packageManager": "yarn@1.22.22+sha256.c17d3797fb9a9115bf375e31bfd30058cac6bc9c3b8807a3d8cb2094794b51ca"
-}
