+++ functions/src/index.ts
@@ -1,0 +1,9 @@
+import { logger } from './config';
+
+// Import and export all functions
+export { home } from './home';
+export { auth } from './auth';
+export { callback } from './callback';
+
+// Log startup
+logger.logOther?.('..:: Alwatr Decap CMS Backend Firebase Functions ::..');
