+++ package.json
@@ -7,19 +7,19 @@
   "author": "S. Amir Mohammad Najafi <njfamirm@gamil.com> (https://www.njfamirm.ir)",
   "type": "module",
   "scripts": {
-    "start": "NODE_OPTIONS=--enable-source-maps run-s clean build serve",
-    "build": "tsc && esbuild src/index.ts --platform=node --external:simple-oauth2 --target=node22 --bundle --format=esm --minify --sourcemap --outdir=dist --out-extension:.js=.mjs",
-    "clean": "rm -rf dist build .tsbuildinfo",
-    "serve": "node --enable-source-maps dist/index.mjs",
-    "watch": "run-s clean build && run-p watch:es watch:node",
-    "watch:node": "nodemon -w dist/ --enable-source-maps dist/index.mjs",
-    "watch:es": "yarn build:es --watch",
+    "start": "firebase emulators:start --only functions",
+    "build": "npm --prefix functions run build",
+    "clean": "rm -rf functions/lib",
+    "serve": "firebase emulators:start --only functions",
+    "deploy": "firebase deploy --only functions",
+    "setup": "firebase functions:config:set github.client_id=\"YOUR_CLIENT_ID\" github.client_secret=\"YOUR_CLIENT_SECRET\" github.scope=\"repo\"",
     "release": "standard-version --sign",
     "format": "prettier . --ignore-path .prettierignore --write"
   },
   "dependencies": {
     "@alwatr/logger": "^5.5.3",
-    "@alwatr/nanotron": "^4.9.1",
+    "firebase-admin": "^11.8.0",
+    "firebase-functions": "^4.3.1",
     "simple-oauth2": "~5.1.0",
     "tslib": "^2.8.1"
   },
@@ -27,8 +27,7 @@
     "@nexim/prettier-config": "^1.3.0",
     "@types/node": "^22.15.3",
     "@types/simple-oauth2": "^5.0.7",
-    "esbuild": "^0.25.3",
-    "npm-run-all": "^4.1.5",
+    "firebase-tools": "^13.0.0",
     "prettier": "^3.5.3",
     "standard-version": "^9.5.0",
     "typescript": "^5.8.3"
