--- prettier.config.cjs
@@ -1,6 +1,0 @@
-/**
- * @type {import('prettier').Config}
- */
-module.exports = {
-  ...require('@nexim/prettier-config'),
-};
