--- src/index.ts
@@ -1,6 +1,0 @@
-import './route/home.js';
-import './route/auth.js';
-import './route/callback.js';
-import { logger } from './config.js';
-
-logger.logOther?.('..:: Alwatr Decap CMS Backend ::..');
