--- src/lib/api-server.ts
@@ -1,5 +1,0 @@
-import { NanotronApiServer } from '@alwatr/nanotron';
-
-import { config } from '../config.js';
-
-export const apiServer = new NanotronApiServer(config.nanoServer);
