--- tsconfig.json
@@ -1,31 +1,0 @@
-{
-  "compilerOptions": {
-    "incremental": true,
-    "noEmit": true,
-    "target": "ES2020",
-    "module": "CommonJS",
-    "lib": ["ES2023", "DOM", "DOM.Iterable"],
-    "resolveJsonModule": true,
-    "strict": true,
-    "noImplicitAny": true,
-    "strictNullChecks": true,
-    "strictFunctionTypes": true,
-    "strictBindCallApply": true,
-    "strictPropertyInitialization": true,
-    "noImplicitThis": true,
-    "alwaysStrict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitOverride": true,
-    "allowSyntheticDefaultImports": true,
-    "sourceMap": true,
-    "inlineSources": true,
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "useDefineForClassFields": false
-  }
-}
